lift import account task data function to context
