+++ src/renderer/contexts/main/DataBackup/default.ts
@@ -8,4 +8,5 @@ export const defaultDataBackupContext: DataBackupContextInterface = {
   importAddressData: () => new Promise(() => {}),
   importEventData: () => new Promise(() => {}),
   importIntervalData: () => new Promise(() => {}),
+  importAccountTaskData: () => new Promise(() => {}),
 };
